Final Project Code/requirements.txt
flask
flask-wtf
wtforms
flask-mail
flask-bootstrap
flask-pymongo
pandas
numpy
flask
gunicorn
matplotlib
cvxopt
